% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ld.R
\name{calculate_average_ld}
\alias{calculate_average_ld}
\title{Calculates the  ld between two alleles}
\usage{
calculate_average_ld(alleles_pos_1, alleles_pos_2)
}
\arguments{
\item{alleles_pos_1}{alleles at locus 1}

\item{alleles_pos_2}{alleles at locus 2}
}
\value{
a list with two entries: LD and r_squared
}
\description{
calculate the average ld between two loci
}
