% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions.R
\name{calculate_dist_junctions}
\alias{calculate_dist_junctions}
\title{collect the full distribution of junctions in the population}
\usage{
calculate_dist_junctions(pop)
}
\arguments{
\item{pop}{object of the class 'population'}
}
\value{
vector with two entries per individual, each indicating the number of
junctions in the respective chromosomes
}
\description{
calculates the distribution of junctions across the population
}
