% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_heterozygosity.R
\name{calculate_heterozygosity}
\alias{calculate_heterozygosity}
\title{Calculate heterozygosity}
\usage{
calculate_heterozygosity(source_pop, locations, progress_bar = FALSE)
}
\arguments{
\item{source_pop}{Population for which to estimate allele frequencies, or a
list of individuals for which to calculate average heterozygosity}

\item{locations}{A vector indicating the locations (in Morgan) of markers for
which to calculate the heterozygosity}

\item{progress_bar}{Displays a progress_bar if TRUE. Default value is TRUE}
}
\value{
A tibble containing the heterozygosities
}
\description{
Calculate the average population level heterozygosity
}
