% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_input_pop}
\alias{check_input_pop}
\title{check input population}
\usage{
check_input_pop(pop)
}
\arguments{
\item{pop}{population, object to be checked}
}
\value{
corrected input population, or error if not supplied
}
\description{
check input population
}
