% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data.R
\name{combine_input_data}
\alias{combine_input_data}
\title{combine sequence data that was previously read from file into a population}
\usage{
combine_input_data(input_data_list, frequencies = NA, pop_size)
}
\arguments{
\item{input_data_list}{list where each entry is the result of
\code{create_input_data}}

\item{frequencies}{frequency of each entry in the list in the starting
population}

\item{pop_size}{intended population size}
}
\value{
the input data entries are combined to one single population that can
be used to seed \code{simulate_admixture_data}. Output is identical to
\code{create_input_data}
}
\description{
Create data in a format that can be used by GenomeAdmixR,
entries are sampled randomly from each input data set, with replacement.
Probability of sampling from each input data set is driven by the input
frequencies, and total number of individuals sampled is driven by pop_size.
}
