% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data.R
\name{create_artificial_genomeadmixr_data}
\alias{create_artificial_genomeadmixr_data}
\title{function to generate artificial genomeadmixr_data}
\usage{
create_artificial_genomeadmixr_data(
  number_of_individuals,
  marker_locations,
  used_nucleotides = 1:4,
  nucleotide_frequencies = NA
)
}
\arguments{
\item{number_of_individuals}{number of individuals}

\item{marker_locations}{location of markers, either in bp or Morgan}

\item{used_nucleotides}{subset or full set of [1/2/3/4] (reflecting a/c/t/g)}

\item{nucleotide_frequencies}{frequencies of the used nucleotides, if not
provided, equal frequencies are assumed.}
}
\value{
genomeadmixr_data object ready for simulate_admixture_data
}
\description{
function to generate artificial genomeadmixr_data
}
