% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findtype}
\alias{findtype}
\title{find local ancestry}
\usage{
findtype(chrom, pos)
}
\arguments{
\item{chrom}{chromosome}

\item{pos}{position where to check for local ancestry}
}
\value{
local ancestry
}
\description{
returns local ancestry at a site
}
