\name{GenomeAdmixR-internal}
\title{Internal GenomeAdmixR functions}
\alias{test_fish_functions}
\alias{create_iso_female_line}
\alias{findtype}
\alias{create_loci_matrix}
\alias{create_pop_class}
\alias{calculate_average_LD}
\alias{calculate_allele_spectrum_cpp}
\alias{calc_allele_frequencies}
\alias{calculate_dist_junctions}
\alias{plot_dist_junctions}
\alias{create_isofemale_line_cpp}
\alias{plot.individual}
\alias{print.population}
\alias{print.individual}
\alias{verify_individual}
\alias{verify_population}
\alias{population_to_vector}
\alias{create_random_markers}
\alias{create_tibble_from_freq_table}
\alias{create_tibble_from_freq_mat}
\alias{simulate_cpp}
\alias{check_input_pop}
\alias{simulate_migration_cpp}

\description{Internal GenomeAdmixR functions}
\details{These are not to be called by the user}
\keyword{internal}