% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data.R
\name{plink_to_genomeadmixr_data}
\alias{plink_to_genomeadmixr_data}
\title{function to convert plink style (ped/map) data to genome_admixr_data}
\usage{
plink_to_genomeadmixr_data(ped_data, map_data, chosen_chromosome)
}
\arguments{
\item{ped_data}{result of read.table(ped_file, header = F)}

\item{map_data}{result of read.table(map_file, header = F)}

\item{chosen_chromosome}{chromosome of choice}
}
\value{
genomeadmixr_data object ready for simulate_admixture_data
}
\description{
function to convert plink style (ped/map) data to genome_admixr_data
}
