% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot.individual}
\alias{plot.individual}
\title{plot the genome of an individual}
\usage{
\method{plot}{individual}(x, cols = NA, ...)
}
\arguments{
\item{x}{object of type individual}

\item{cols}{colors for the different ancestors}

\item{...}{other arguments}
}
\value{
No return value
}
\description{
visualise ancestry blocks on both chromosomes
}
