% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chromosome.R
\name{plot_chromosome}
\alias{plot_chromosome}
\title{plots a chromosome}
\usage{
plot_chromosome(chrom, xmin = 0, xmax = 1)
}
\arguments{
\item{chrom}{object of type chromosome, typically a table with two columns.
The first column indicates the start of an ancestry block (location in
Morgan), the second column indicates the ancestry type.}

\item{xmin}{minimum value of the range, default = 0.}

\item{xmax}{maximum value of the range, default = 1.}
}
\description{
This function plots a chromosome in the range [xmin, xmax].
Colors indicate different ancestry.
}
