% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_difference_frequencies.R
\name{plot_difference_frequencies}
\alias{plot_difference_frequencies}
\title{Plot the change in frequency between the start and end of a simulation}
\usage{
plot_difference_frequencies(
  results,
  picked_ancestor = "ALL",
  picked_population = 1
)
}
\arguments{
\item{results}{An object which is the result of \code{select_population} or
\code{create_population_selection}, being a list with four properties:
\code{population}, \code{frequencies}, \code{initial_frequencies} and
\code{final frequencies}}

\item{picked_ancestor}{Default is "ALL", where different colors indicate
different ancestors. Alternatively, for clarity, the user can specify a
specific ancestral allele, and only that allele is plotted}

\item{picked_population}{If multiple populations were simulated (in the case
of \code{simulate_admixture_migration}), which population should be plotted?
Default is population_1.}
}
\value{
a ggplot2 object
}
\description{
This function plots the change in frequency of one or
multiple ancestors after performing a simulation.
}
