% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions.R
\name{plot_dist_junctions}
\alias{plot_dist_junctions}
\title{plot the distribution of junctions}
\usage{
plot_dist_junctions(pop)
}
\arguments{
\item{pop}{of the class 'population'}
}
\value{
No return value
}
\description{
plots the distribution of junctions in the population using
base R
}
