% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.genomeadmixr_data}
\alias{print.genomeadmixr_data}
\title{print an individual to the console}
\usage{
\method{print}{genomeadmixr_data}(x, ...)
}
\arguments{
\item{x}{individual}

\item{...}{other arguments}
}
\description{
prints an object of class genomeadmixr_data to the console
}
