% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.individual}
\alias{print.individual}
\title{print an individual to the console}
\usage{
\method{print}{individual}(x, ...)
}
\arguments{
\item{x}{individual}

\item{...}{other arguments}
}
\value{
No return value
}
\description{
prints an object of class individual to the console
}
