% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.population}
\alias{print.population}
\title{print a population object}
\usage{
\method{print}{population}(x, ...)
}
\arguments{
\item{x}{input population}

\item{...}{other arguments}
}
\value{
No return value
}
\description{
prints the contents of a population nicely
}
