% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{save_population}
\alias{save_population}
\title{Save a population to file}
\usage{
save_population(population, file_name, compression = TRUE)
}
\arguments{
\item{population}{Object of class \code{population}}

\item{file_name}{Name of the file to save the population}

\item{compression}{By default, the population is compressed to reduce file
size. See for more information \code{saveRDS}}
}
\description{
Saves a population to file for later use
}
\details{
This function functions as a wrapper for the base function
\code{saveRDS}.
}
\examples{
wildpop <- simulate_admixture(pop_size = 10,
number_of_founders = 2,
total_runtime = 3,
morgan = 1,
seed = 123)
save_population(wildpop, file_name = "wildpop.pop")
}
