% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data.R
\name{simulation_data_to_genomeadmixr_data}
\alias{simulation_data_to_genomeadmixr_data}
\title{function to convert ped/map data to genome_admixr_data}
\usage{
simulation_data_to_genomeadmixr_data(
  simulation_data,
  markers = NA,
  verbose = FALSE
)
}
\arguments{
\item{simulation_data}{result of simulate_admixture}

\item{markers}{vector of locations of markers (in Morgan). If no vector is
provided, the function searches for marker locations in the simulation_data.}

\item{verbose}{provide verbose output (default is FALSE)}
}
\value{
genomeadmixr_data object ready for simulate_admixture_data
}
\description{
function to convert ped/map data to genome_admixr_data
}
