% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulation_data_to_genomeadmixr_data_cpp}
\alias{simulation_data_to_genomeadmixr_data_cpp}
\title{function to convert data}
\usage{
simulation_data_to_genomeadmixr_data_cpp(input_population, markers)
}
\arguments{
\item{input_population}{input population}

\item{markers}{markers}
}
\description{
function to convert data
}
