% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data.R
\name{vcfR_to_genomeadmixr_data}
\alias{vcfR_to_genomeadmixr_data}
\title{function to convert a vcfR object to genome_admixr_data}
\usage{
vcfR_to_genomeadmixr_data(vcfr_object, chosen_chromosome)
}
\arguments{
\item{vcfr_object}{result of vcfR::read.vcfR}

\item{chosen_chromosome}{chromosome of choice}
}
\value{
genomeadmixr_data object ready for simulate_admixture_data
}
\description{
function to convert a vcfR object to genome_admixr_data
}
