% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{verify_individual}
\alias{verify_individual}
\title{verify that an individual is correct}
\usage{
verify_individual(indiv)
}
\arguments{
\item{indiv}{object of class 'individual'}
}
\description{
function to verify correctness of an object of class 'individual'
}
