% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{verify_population}
\alias{verify_population}
\title{verify that a population is correct}
\usage{
verify_population(pop)
}
\arguments{
\item{pop}{object of class 'population'}
}
\description{
function to verify correctness of an object of class 'population'
}
