% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data.R
\name{write_as_plink}
\alias{write_as_plink}
\title{function to write simulation output as PLINK style data}
\usage{
write_as_plink(
  input_pop,
  marker_locations,
  file_name_prefix,
  chromosome = 1,
  recombination_rate = 1
)
}
\arguments{
\item{input_pop}{input population, either of class "population" or of class
"genomeadmixr_data"}

\item{marker_locations}{location of markers, in bp}

\item{file_name_prefix}{prefix of the ped/map files.}

\item{chromosome}{chromosome indication for map file}

\item{recombination_rate}{recombination rate in cM / kb}
}
\value{
No return value
}
\description{
function to write simulation output as PLINK style data
}
