% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{load_population}
\alias{load_population}
\title{Load a population from file}
\usage{
load_population(file_name)
}
\arguments{
\item{file_name}{Name of the file to save the population}
}
\value{
A population object
}
\description{
Loads a population that has previously been written to file.
}
\details{
This function is a wrapper for \code{readRDS}.
}
\examples{
wildpop <- simulate_admixture(pop_size = 10,
number_of_founders = 2,
total_runtime = 3,
morgan = 1,
seed = 123)
save_population(wildpop, file_name = "wildpop.pop")
wildpop2 <- load_population(file_name = "wildpop.pop")
all.equal(wildpop, wildpop2)
}
\seealso{
\code{\link{save_population}}
}
